Add shane's first data scrape
Shane scraped this data and gave it to us on Friday 11/11/16.  Below is his message:

"here are the per game career stats of every player currently on an nba roster
this is just to give you guys something to work off til i get more. but from here i plan append to each row here the career info for each player from the 'Per 100 Poss', 'Advanced', and 'Shooting'  tables
my idea from there would be to do the same thing except instead of just a career entry for each player, have an entry for each season a player played. so, 2011Lebron would be a different row than 2014Lebron and both different then CareerLebron."